+++ src/providers/next-auth-providers.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { type ReactNode } from "react";
+
+import { SessionProvider } from "next-auth/react";
+
+export default function NextAuthProvider({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return <SessionProvider>{children}</SessionProvider>;
+}
