+++ src/utils/next-auth-options.ts
@@ -18,9 +18,6 @@ export const nextAuthOptions: NextAuthOptions = {
   ],
   callbacks: {
     jwt: async ({ token, user, account, profile }) => {
-      // 注意: トークンをログ出力してはダメです。
-      console.log("in jwt", { user, token, account, profile });
-
       if (user) {
         token.user = user;
         const u = user as any;
